Create inpage error page for iframes
- Split out the error work so it can be used in multi places
- Create a minimal page template which is a page without header and
  footer
