+++ src/server/common/helpers/errors/catch-all.js
@@ -1,17 +1,4 @@
-function statusCodeMessage(statusCode) {
-  switch (true) {
-    case statusCode === 404:
-      return 'Page not found'
-    case statusCode === 403:
-      return 'Forbidden'
-    case statusCode === 401:
-      return 'Unauthorized'
-    case statusCode === 400:
-      return 'Bad Request'
-    default:
-      return 'Something went wrong'
-  }
-}
+import { statusCodeMessage } from '~/src/server/common/helpers/errors/status-code-message'
 
 function catchAll(request, h) {
   const { response } = request
