+++ src/server/common/helpers/errors/status-code-message.js
@@ -1,0 +1,16 @@
+function statusCodeMessage(statusCode) {
+  switch (true) {
+    case statusCode === 404:
+      return 'Page not found'
+    case statusCode === 403:
+      return 'Forbidden'
+    case statusCode === 401:
+      return 'Unauthorized'
+    case statusCode === 400:
+      return 'Bad Request'
+    default:
+      return 'Something went wrong'
+  }
+}
+
+export { statusCodeMessage }
