+++ src/server/error/minimal-error.njk
@@ -1,0 +1,6 @@
+{# Minimal error page a error page without header or footer #}
+{% extends "layouts/minimal-page.njk" %}
+
+{% block content %}
+  {% include "error/partials/error.njk" %}
+{% endblock %}
