+++ src/server/index.js
@@ -7,7 +7,7 @@ import { router } from '~/src/server/router'
 import { config } from '~/src/config'
 import { nunjucksConfig } from '~/src/config/nunjucks'
 import { csrf } from '~/src/server/common/helpers/auth/csrf'
-import { catchAll } from '~/src/server/common/helpers/errors'
+import { catchAll } from '~/src/server/common/helpers/errors/catch-all'
 import { azureOidc } from '~/src/server/common/helpers/auth/azure-oidc'
 import { buildRedisClient } from '~/src/server/common/helpers/redis-client'
 import { sessionManager } from '~/src/server/common/helpers/session-manager'
